+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Enhancements
 - Added new `stat_type` for UncrossingPrice.
+- Added new publisher values for `XNAS.BASIC`
+
+### Bug fixes
+- Fix descriptions for `FINN` and `FINY` publishers.
 
 ## 0.18.0 - 2024-05-21